Wordle solver and the two word lists as text files
This is the Worldle solver in Python, and Wordle's two word lists as text files. Wordle uses both an 'allowed guesses' word list, and an 'answers' list. Together there are over 12k words on these lists. My program uses both word lists, because if I make it too easy, it's not fun anymore!